back/requirements.txt
fastapi==0.98.0
uvicorn==0.22.0
yoyo-migrations==8.2.0
SQLAlchemy==2.0.17
psycopg2==2.9.6
PyJWT==2.7.0
python-multipart~=0.0.6